The Colletions Explorer now works properly for non-superusers.
Previously, the code had not properly granted viewing permission for the
Collections Explorer to users who had the appropriate permission. This was
because get_collections_with_management_permission() was overloaded to do
something it had no business doing, and it was buggy.

Now, it no longer deals with parent collections; the sole part of the code that
uses that argument now just returns the children of the parent directly. This
means that management permissions are not improperly checked for children of
already permitted Collections, so users can actually see and edit the children
of permitted Collections.